+++ packages/moonstone/CHANGELOG.md
@@ -38,6 +38,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VideoPlayer` to correctly change sources
 - `moonstone/VideoPlayer` to show or hide feedback tooltip properly
 - `moonstone/DateTimeDecorator` to work properly with `RadioControllerDecorator`
+- `moonstone/Marquee` to correctly show disabled state
 
 ### Removed